requirements.txt
easygui==0.98.1
lxml==4.6.1
pkg-resources==0.0.0
python-docx==0.8.10
